// textbook/content/shared/types.d.ts
export class Blank extends CustomElementView {
  private $target;
  private solution;
  done: boolean;
  ready(): void;
  solve(restore?: boolean): void;
}

export class BlankInput extends CustomElementView {
  private $input;
  private $target;
  private solution;
  private solutionNum;
  private range;
  private input;
  private attempts;
  private placeholder;
  done: boolean;
  ready(): void;
  get isCorrect(): boolean;
  moveCursor(): void;
  solve(restore?: boolean): void;
  focus(): void;
  blur(): void;
}

export type EquationValidationResponse = {isCorrect?: boolean; error?: string}|undefined;

export class Equation extends CustomElementView {
  $math: ElementView;
  private $textarea;
  private $cursor;
  private $error;
  private $step?;
  private numeric?;
  private vars;
  private isDone;
  private shortVar;
  private value?;
  private error?;
  private lastAttempt;
  private errorCount;
  private autocomplete?;
  private substitutions?;
  solution: ExprElement;
  hints: string[];
  fns: string[];
  validate?: ((expr: ExprElement) => EquationValidationResponse);
  ready(): void;
  focus(): void;
  onPointerdown(e: PointerEvent): void;
  onKey(e: {code?: number; key: string}): void;
  onBlur(): void;
  check(): string|void;
  complete(expr: ExprElement): void;
  solve(): void;
}

export class EquationSystem extends CustomElementView {
  private $table;
  private $equation;
  private lastRowContent;
  private rowCount;
  private previousAnswers;
  private isSolved;
  private maxRows;
  private steps;
  private hints;
  validate?: (expr: ExprElement, isRepeated: boolean) => EquationValidationResponse;
  isFinal?: (expr: ExprElement) => boolean;
  created(): void;
  ready(): void;
  setupEquation($equation: Equation): void;
  onSolveRow(expr: ExprElement): void;
  solve(): void;
}

export class Picker extends CustomElementView {
  private $items;
  private itemsClicked;
  private correctCount;
  private solvedCount;
  private isSolved;
  ready(): void;
  restore(scores: string[]): void;
  checkSolved(): void;
}

export class Sortable extends CustomElementView {
  private items;
  ready(): void;
  solve(): void;
}

export interface HintOptions {
  class?: string;
  visible?: boolean;
  store?: boolean;
  force?: boolean;
  variables?: Obj<string>;
  timeout?: number;
  toast?: boolean;
}
// ...
export class Step extends CustomElementView {
  private $course?;
  model: Observable;
  tools: {confetti: (duration?: number, maxParticles?: number) => void};
  initialData?: {scores: string[]; data: Obj<any>};
  isShown: boolean;
  isCompleted: boolean;
  goals: string[];
  scores: Set<string>;
  $blanks: (Blank|BlankInput)[];
  $equations: (Equation|EquationSystem)[];
  $reveals: ElementView[];
  $nextBtn: ElementView[];
  $picker?: Picker;
  $sortables: Sortable[];
  ready(): void;
  show(): void;
  complete(): void;
  restore(userData: any): void;
  get isReady(): boolean;
  get isPageLoaded(): boolean;
  score(goal: string, goNext?: boolean): void;
  storeData(key: string, value: any): void;
  onScore(goalList: string, callback?: () => void): Promise<void>;
  addHint(text: string, options?: HintOptions): {text: string};
  getText(id: string): string;
  getHelp(): void;
  delayedHint(callback: () => void, t?: number): void;
}